requirements.txt
numpy>=1.21.0
scipy>=1.7.0
biopython>=1.79
matplotlib>=3.4.0